refactor code to function based structure to avoid circular imports
