+++ src/store/actions.ts
@@ -1,0 +1,3 @@
+import { createAction } from '@reduxjs/toolkit';
+
+export const resetApp = createAction('app/reset');
