+++ src/store/slices/moviesSlice.ts
@@ -1,0 +1,24 @@
+import { createSlice } from '@reduxjs/toolkit';
+import { resetApp } from './../index';
+
+const moviesSlice = createSlice({
+  name: 'movies',
+  initialState: [] as string[],
+  reducers: {
+    addMovie(state, action) {
+      state.push(action.payload);
+    },
+    removeMovie(state, action) {
+      const index = state.indexOf(action.payload);
+      state.splice(index, 1);
+    },
+  },
+  extraReducers(builder) {
+    builder.addCase(resetApp.toString(), (/* state, action */) => {
+      return [];
+    });
+  },
+});
+
+export const { addMovie, removeMovie } = moviesSlice.actions;
+export const moviesReducer = moviesSlice.reducer;
